+++ app/containers/Vartan/render.js
@@ -28,18 +28,19 @@ const render = ({ regl, citiesData, imgData, width, height }) => {
   const toSwarm = points => swarmLayout(points, width, height, citiesData);
   const toPhoto = points => photoLayout(points, width, height, imgData);
   const toArea = points => areaLayout(points, width, height, citiesData);
-  const toPhyllotaxis = points => phyllotaxisLayout(
-    points,
-    pointWidth,
-    width / 2,
-    height / 2,
-    citiesData
-  );
-  const toBlack = points => points.map(d =>
-    d.color = [0, 0, 0]
-  );
+  const toPhyllotaxis = points =>
+    phyllotaxisLayout(points, pointWidth, width / 2, height / 2, citiesData);
+  const toBlack = points => points.map(d => (d.color = [0, 0, 0]));
 
-  const layouts = [toPhyllotaxis, toSwarm, toCities, toArea, toBars, toPhoto, toBlack];
+  const layouts = [
+    toPhyllotaxis,
+    toSwarm,
+    toCities,
+    toArea,
+    toBars,
+    toPhoto,
+    toBlack,
+  ];
   let currentLayout = 0;
 
   // wrap d3 color scales so they produce vec3s with values 0-1
@@ -50,7 +51,7 @@ const render = ({ regl, citiesData, imgData, width, height }) => {
       .range([0.4, 1]);
     const _rgb = rgb(scale(tScale(t)));
     return [_rgb.r / 255, _rgb.g / 255, _rgb.b / 255];
-  }
+  };
 
   const colorScales = [
     scaleSequential(interpolateViridis),
@@ -61,46 +62,47 @@ const render = ({ regl, citiesData, imgData, width, height }) => {
   let currentColorScale = 0;
 
   // function to compile a draw points regl func
-  const createDrawPoints = points => regl({
-    frag:
-      '\n\t\t  precision highp float;\n\t\t\tvarying vec3 fragColor;\n\t\t\tvoid main() {\n\t\t\t\tgl_FragColor = vec4(fragColor, 1);\n\t\t\t}\n\t\t\t',
+  const createDrawPoints = points =>
+    regl({
+      frag:
+        '\n\t\t  precision highp float;\n\t\t\tvarying vec3 fragColor;\n\t\t\tvoid main() {\n\t\t\t\tgl_FragColor = vec4(fragColor, 1);\n\t\t\t}\n\t\t\t',
 
-    vert:
-      '\n\t\t\tattribute vec2 positionStart;\n\t\t\tattribute vec2 positionEnd;\n\t\t\tattribute float index;\n\t\t\tattribute vec3 colorStart;\n\t\t\tattribute vec3 colorEnd;\n\n\t\t\tvarying vec3 fragColor;\n\n\t\t\tuniform float pointWidth;\n\t\t\tuniform float stageWidth;\n\t\t\tuniform float stageHeight;\n\t\t\tuniform float elapsed;\n\t\t\tuniform float duration;\n\t\t\tuniform float delayByIndex;\n\t\t\t// uniform float tick;\n\t\t\t// uniform float animationRadius;\n\t\t\tuniform float numPoints;\n\n\t\t\t// helper function to transform from pixel space to normalized device coordinates (NDC)\n\t\t\t// in NDC (0,0) is the middle, (-1, 1) is the top left and (1, -1) is the bottom right.\n\t\t\tvec2 normalizeCoords(vec2 position) {\n\t\t\t\t// read in the positions into x and y vars\n\t      float x = position[0];\n\t      float y = position[1];\n\n\t\t\t\treturn vec2(\n\t\t      2.0 * ((x / stageWidth) - 0.5),\n\t\t      // invert y since we think [0,0] is bottom left in pixel space\n\t\t      -(2.0 * ((y / stageHeight) - 0.5)));\n\t\t\t}\n\n\t\t\t// helper function to handle cubic easing (copied from d3 for consistency)\n\t\t\t// note there are pre-made easing functions available via glslify.\n\t\t\tfloat easeCubicInOut(float t) {\n\t\t\t\tt *= 2.0;\n        t = (t <= 1.0 ? t * t * t : (t -= 2.0) * t * t + 2.0) / 2.0;\n\n        if (t > 1.0) {\n          t = 1.0;\n        }\n\n        return t;\n\t\t\t}\n\n\t\t\tvoid main() {\n\t\t\t\tgl_PointSize = pointWidth;\n\n\t\t\t\tfloat delay = delayByIndex * index;\n\t      float t;\n\n\t      // drawing without animation, so show end state immediately\n\t      if (duration == 0.0) {\n\t        t = 1.0;\n\n\t      // still delaying before animating\n\t      } else if (elapsed < delay) {\n\t        t = 0.0;\n\t      } else {\n\t        t = easeCubicInOut((elapsed - delay) / duration);\n\t      }\n\n\t      // interpolate position\n\t      vec2 position = mix(positionStart, positionEnd, t);\n\n\t      // apply an ambient animation\n\t\t\t\t// float dir = index > numPoints / 2.0 ? 1.0 : -1.0;\n\t      // position[0] += animationRadius * cos((tick + index) * dir);\n\t      // position[1] += animationRadius * sin((tick + index) * dir);\n\n\t      // above we + index to offset how they move\n\t      // we multiply by dir to change CW vs CCW for half\n\n\n\t      // interpolate color\n\t      fragColor = mix(colorStart, colorEnd, t);\n\n\t      // scale to normalized device coordinates\n\t\t\t\t// gl_Position is a special variable that holds the position of a vertex\n\t      gl_Position = vec4(normalizeCoords(position), 0.0, 1.0);\n\t\t\t}\n\t\t\t',
+      vert:
+        '\n\t\t\tattribute vec2 positionStart;\n\t\t\tattribute vec2 positionEnd;\n\t\t\tattribute float index;\n\t\t\tattribute vec3 colorStart;\n\t\t\tattribute vec3 colorEnd;\n\n\t\t\tvarying vec3 fragColor;\n\n\t\t\tuniform float pointWidth;\n\t\t\tuniform float stageWidth;\n\t\t\tuniform float stageHeight;\n\t\t\tuniform float elapsed;\n\t\t\tuniform float duration;\n\t\t\tuniform float delayByIndex;\n\t\t\t// uniform float tick;\n\t\t\t// uniform float animationRadius;\n\t\t\tuniform float numPoints;\n\n\t\t\t// helper function to transform from pixel space to normalized device coordinates (NDC)\n\t\t\t// in NDC (0,0) is the middle, (-1, 1) is the top left and (1, -1) is the bottom right.\n\t\t\tvec2 normalizeCoords(vec2 position) {\n\t\t\t\t// read in the positions into x and y vars\n\t      float x = position[0];\n\t      float y = position[1];\n\n\t\t\t\treturn vec2(\n\t\t      2.0 * ((x / stageWidth) - 0.5),\n\t\t      // invert y since we think [0,0] is bottom left in pixel space\n\t\t      -(2.0 * ((y / stageHeight) - 0.5)));\n\t\t\t}\n\n\t\t\t// helper function to handle cubic easing (copied from d3 for consistency)\n\t\t\t// note there are pre-made easing functions available via glslify.\n\t\t\tfloat easeCubicInOut(float t) {\n\t\t\t\tt *= 2.0;\n        t = (t <= 1.0 ? t * t * t : (t -= 2.0) * t * t + 2.0) / 2.0;\n\n        if (t > 1.0) {\n          t = 1.0;\n        }\n\n        return t;\n\t\t\t}\n\n\t\t\tvoid main() {\n\t\t\t\tgl_PointSize = pointWidth;\n\n\t\t\t\tfloat delay = delayByIndex * index;\n\t      float t;\n\n\t      // drawing without animation, so show end state immediately\n\t      if (duration == 0.0) {\n\t        t = 1.0;\n\n\t      // still delaying before animating\n\t      } else if (elapsed < delay) {\n\t        t = 0.0;\n\t      } else {\n\t        t = easeCubicInOut((elapsed - delay) / duration);\n\t      }\n\n\t      // interpolate position\n\t      vec2 position = mix(positionStart, positionEnd, t);\n\n\t      // apply an ambient animation\n\t\t\t\t// float dir = index > numPoints / 2.0 ? 1.0 : -1.0;\n\t      // position[0] += animationRadius * cos((tick + index) * dir);\n\t      // position[1] += animationRadius * sin((tick + index) * dir);\n\n\t      // above we + index to offset how they move\n\t      // we multiply by dir to change CW vs CCW for half\n\n\n\t      // interpolate color\n\t      fragColor = mix(colorStart, colorEnd, t);\n\n\t      // scale to normalized device coordinates\n\t\t\t\t// gl_Position is a special variable that holds the position of a vertex\n\t      gl_Position = vec4(normalizeCoords(position), 0.0, 1.0);\n\t\t\t}\n\t\t\t',
 
-    attributes: {
-      positionStart: points.map(d => [d.sx, d.sy]),
-      positionEnd: points.map(d => [d.tx, d.ty]),
-      colorStart: points.map(d => d.colorStart),
-      colorEnd: points.map(d => d.colorEnd),
-      index: range(points.length),
-    },
-
-    uniforms: {
-      pointWidth: regl.prop('pointWidth'),
-      stageWidth: regl.prop('stageWidth'),
-      stageHeight: regl.prop('stageHeight'),
-      delayByIndex: regl.prop('delayByIndex'),
-      duration: regl.prop('duration'),
-      numPoints: numPoints,
-      // animationRadius: 0,// 15.0,
-      // tick: (reglprops) => { // increase multiplier for faster animation speed
-      // 	// console.log(reglprops);
-      // 	// return reglprops.tick / 50;
-      // 	return 0; // disable ambient animation
-      // },
-      // time in milliseconds since the prop startTime (i.e. time elapsed)
-      elapsed: (_ref, _ref2) => {
-        const time = _ref.time;
-        const _ref2$startTime = _ref2.startTime,
-          startTime = _ref2$startTime === undefined ? 0 : _ref2$startTime;
-        return (time - startTime) * 1000;
+      attributes: {
+        positionStart: points.map(d => [d.sx, d.sy]),
+        positionEnd: points.map(d => [d.tx, d.ty]),
+        colorStart: points.map(d => d.colorStart),
+        colorEnd: points.map(d => d.colorEnd),
+        index: range(points.length),
       },
-    },
 
-    count: points.length,
-    primitive: 'points',
-  });
+      uniforms: {
+        pointWidth: regl.prop('pointWidth'),
+        stageWidth: regl.prop('stageWidth'),
+        stageHeight: regl.prop('stageHeight'),
+        delayByIndex: regl.prop('delayByIndex'),
+        duration: regl.prop('duration'),
+        numPoints: numPoints,
+        // animationRadius: 0,// 15.0,
+        // tick: (reglprops) => { // increase multiplier for faster animation speed
+        // 	// console.log(reglprops);
+        // 	// return reglprops.tick / 50;
+        // 	return 0; // disable ambient animation
+        // },
+        // time in milliseconds since the prop startTime (i.e. time elapsed)
+        elapsed: (_ref, _ref2) => {
+          const time = _ref.time;
+          const _ref2$startTime = _ref2.startTime,
+            startTime = _ref2$startTime === undefined ? 0 : _ref2$startTime;
+          return (time - startTime) * 1000;
+        },
+      },
+
+      count: points.length,
+      primitive: 'points',
+    });
 
   // function to start animation loop (note: time is in seconds)
   function animate(layout, points) {
@@ -183,9 +185,9 @@ const render = ({ regl, citiesData, imgData, width, height }) => {
 
   // create initial set of points
   const points = range(numPoints).map(d => ({
-    tx : width / 2,
-    ty : height / 2,
-    colorEnd: [0, 0, 0]
+    tx: width / 2,
+    ty: height / 2,
+    colorEnd: [0, 0, 0],
   }));
 
   // start animation loop
